+++ craft-ui/src/components/stats_modal.rs
@@ -110,7 +110,7 @@ impl StatsModal {
         frame.render_widget(paragraph, inner);
 
         if total > viewport_h {
-            render_vertical_scrollbar(frame, inner, total, scroll);
+            render_vertical_scrollbar(frame, inner, u32::from(total), u32::from(scroll));
         }
 
         popup
